refactor: app deploy
