+++ .github/workflows/app-deploy.yml
@@ -120,12 +120,6 @@ jobs:
           state: ${{ job.status }}
           environment-url: ${{ steps.deployment-url.outputs.deployment-url }}
           deployment-id: ${{ steps.deployment.outputs.deployment_id }}
-      - name: Install Playwright dependencies
-        run: npx playwright install-deps
-      - name: Run Lighthouse tests
-        run: npx nx run-many --target=lighthouse --projects=${{ matrix.app }}
-        env:
-          DEPLOYMENT_URL: ${{ steps.deployment-url.outputs.deployment-url }}
   deploy-production:
     name: Deploy Production
     # main branch and not a pull request
@@ -198,9 +192,3 @@ jobs:
           state: ${{ job.status }}
           environment-url: ${{ steps.deployment-url.outputs.deployment-url }}
           deployment-id: ${{ steps.deployment.outputs.deployment_id }}
-      - name: Install Playwright dependencies
-        run: npx playwright install-deps
-      - name: Run Lighthouse tests
-        run: npx nx run-many --target=lighthouse --projects=${{ matrix.app }}
-        env:
-          DEPLOYMENT_URL: ${{ steps.deployment-url.outputs.deployment-url }}
